Parse HWP errors
RTC:151336
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/27340
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/usr/fapi2/plat_utils.C
@@ -809,7 +809,7 @@ errlHndl_t rcToErrl(ReturnCode & io_rc,
 
     if (io_rc)
     {
-        uint64_t l_rcValue = io_rc;
+        uint32_t l_rcValue = io_rc;
 
         // ReturnCode contains an error. Find out which component of the HWPF
         // created the error
